chore: move email module to pkg
