+++ pkg/email/templates/password_reset.html
